Rename utils to common, fix back buttons for menu pages
