+++ tabs/menu/profile.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var Back = require('../../components/utils').BackButton;
+var Back = require('../../components/common').BackButton;
 
 var React = require('react-native');
 var {
